+++ src/ezsift.cpp
@@ -24,10 +24,7 @@
 */
 
 #include "ezsift.h"
-
-#include "common.h"
 #include "image.h"
-#include "image_utility.h"
 #include "timer.h"
 #include "vvector.h"
 
@@ -144,6 +141,7 @@ int row_filter_transpose(float *src, float *dst, int w, int h, float *coef1d,
         }
     }
     delete[] row_buf;
+    row_buf = nullptr;
 
     return 0;
 }
@@ -428,6 +426,7 @@ float compute_orientation_hist(const Image<float> &image, SiftKeypoint &kpt,
     kpt.ori = max_i * _2PI / nBins;
 
     delete[] tmpHist;
+    tmpHist = nullptr;
     return maxitem;
 }
 
@@ -555,6 +554,7 @@ float compute_orientation_hist_with_gradient(const Image<float> &grdImage,
     kpt.ori = max_i * _2PI / nBins;
 
     delete[] tmpHist;
+    tmpHist = nullptr;
     return maxitem;
 }
 
@@ -718,6 +718,7 @@ int detect_keypoints(std::vector<Image<float>> &dogPyr,
     }
 
     delete[] hist;
+    hist = nullptr;
     return 0;
 }
 
@@ -909,13 +910,11 @@ int extract_descriptor(std::vector<Image<float>> &grdPyr,
     int nBins = nSubregion * nSubregion * nBinsPerSubregion;
     int nHistBins =
         (nSubregion + 2) * (nSubregion + 2) * (nBinsPerSubregion + 2);
-    int nSliceStep = (nSubregion + 2) * (nBinsPerSubregion + 2); // 32
+    int nSliceStep = (nSubregion + 2) * (nBinsPerSubregion + 2);
     int nRowStep = (nBinsPerSubregion + 2);
     float *histBin = new float[nHistBins];
 
-    float exp_scale =
-        -2.0f /
-        (nSubregion * nSubregion); // -1/(2* nSubregion/2 * nSubregion/2)
+    float exp_scale = -2.0f / (nSubregion * nSubregion);
 
     for (std::list<SiftKeypoint>::iterator kpt = kpt_list.begin();
          kpt != kpt_list.end(); kpt++) {
@@ -1139,12 +1138,16 @@ int extract_descriptor(std::vector<Image<float>> &grdPyr,
 
         memcpy(kpt->descriptors, dstBins, nBins * sizeof(float));
 
-        if (dstBins)
+        if (dstBins) {
             delete[] dstBins;
+            dstBins = nullptr;
+        }
     }
 
-    if (histBin)
+    if (histBin) {
         delete[] histBin;
+        histBin = nullptr;
+    }
 
     return 0;
 }
@@ -1220,299 +1223,6 @@ int sift_cpu(const Image<unsigned char> &image,
     if (bExtractDescriptors)
         extract_descriptor(grdPyr, rotPyr, nOctaves, nGpyrLayers, kpt_list);
 
-    return 0;
-}
-
-//////////////////////
-// Helper Functions //
-//////////////////////
-
-// Combine two images horizontally
-int combine_image(Image<unsigned char> &out_image,
-                  const Image<unsigned char> &image1,
-                  const Image<unsigned char> &image2)
-{
-    int w1 = image1.w;
-    int h1 = image1.h;
-    int w2 = image2.w;
-    int h2 = image2.h;
-    int dstW = w1 + w2;
-    int dstH = MAX(h1, h2);
-
-    out_image.init(dstW, dstH);
-
-    unsigned char *srcData1 = image1.data;
-    unsigned char *srcData2 = image2.data;
-    unsigned char *dstData = out_image.data;
-
-    for (int r = 0; r < dstH; r++) {
-        if (r < h1) {
-            memcpy(dstData, srcData1, w1 * sizeof(unsigned char));
-        }
-        else {
-            memset(dstData, 0, w1 * sizeof(unsigned char));
-        }
-        dstData += w1;
-
-        if (r < h2) {
-            memcpy(dstData, srcData2, w2 * sizeof(unsigned char));
-        }
-        else {
-            memset(dstData, 0, w2 * sizeof(unsigned char));
-        }
-        dstData += w2;
-        srcData1 += w1;
-        srcData2 += w2;
-    }
-
-    return 0;
-}
-
-// Helper callback function for merge match list.
-bool same_match_pair(MatchPair first, MatchPair second)
-{
-    if (first.c1 == second.c1 && first.r1 == second.r1 &&
-        first.c2 == second.c2 && first.r2 == second.r2)
-        return true;
-    else
-        return false;
-}
-
-// Match keypoints from two images, using brutal force method.
-// Use Euclidean distance as matching score.
-int match_keypoints(std::list<SiftKeypoint> &kpt_list1,
-                    std::list<SiftKeypoint> &kpt_list2,
-                    std::list<MatchPair> &match_list)
-{
-    std::list<SiftKeypoint>::iterator kpt1;
-    std::list<SiftKeypoint>::iterator kpt2;
-
-    for (kpt1 = kpt_list1.begin(); kpt1 != kpt_list1.end(); kpt1++) {
-        // Position of the matched feature.
-        int r1 = (int)kpt1->r;
-        int c1 = (int)kpt1->c;
-
-        float *descr1 = kpt1->descriptors;
-        float score1 = (std::numeric_limits<float>::max)(); // highest score
-        float score2 = (std::numeric_limits<float>::max)(); // 2nd highest score
-
-        // Position of the matched feature.
-        int r2 = 0, c2 = 0;
-        for (kpt2 = kpt_list2.begin(); kpt2 != kpt_list2.end(); kpt2++) {
-            float score = 0;
-            float *descr2 = kpt2->descriptors;
-            float dif;
-            for (int i = 0; i < DEGREE_OF_DESCRIPTORS; i++) {
-                dif = descr1[i] - descr2[i];
-                score += dif * dif;
-            }
-
-            if (score < score1) {
-                score2 = score1;
-                score1 = score;
-                r2 = (int)kpt2->r;
-                c2 = (int)kpt2->c;
-            }
-            else if (score < score2) {
-                score2 = score;
-            }
-        }
-
-#if (USE_FAST_FUNC == 1)
-        if (fast_sqrt_f(score1 / score2) < SIFT_MATCH_NNDR_THR)
-#else
-        if (sqrtf(score1 / score2) < SIFT_MATCH_NNDR_THR)
-#endif
-        {
-            MatchPair mp;
-            mp.r1 = r1;
-            mp.c1 = c1;
-            mp.r2 = r2;
-            mp.c2 = c2;
-
-            match_list.push_back(mp);
-        }
-    }
-
-    match_list.unique(same_match_pair);
-
-#if PRINT_MATCH_KEYPOINTS
-    std::list<MatchPair>::iterator p;
-    int match_idx = 0;
-    for (p = match_list.begin(); p != match_list.end(); p++) {
-        printf("\tMatch %3d: (%4d, %4d) -> (%4d, %4d)\n", match_idx, p->r1,
-               p->c1, p->r2, p->c2);
-        match_idx++;
-    }
-#endif
-
-    return 0;
-}
-
-void draw_keypoints_to_ppm_file(const char *out_filename,
-                                const Image<unsigned char> &image,
-                                std::list<SiftKeypoint> kpt_list)
-{
-    std::list<SiftKeypoint>::iterator it;
-    ImagePPM imgPPM;
-    int w = image.w;
-    int h = image.h;
-    int r, c;
-
-    /*******************************
-     * cR:
-     * radius of the circle
-     * cR = sigma * 4 * (2^O)
-     *******************************/
-    int cR;
-
-    // initialize the imgPPM
-    imgPPM.w = w;
-    imgPPM.h = h;
-    imgPPM.img_r = new unsigned char[w * h];
-    imgPPM.img_g = new unsigned char[w * h];
-    imgPPM.img_b = new unsigned char[w * h];
-
-    int i, j;
-    unsigned char *data = image.data;
-    // Copy gray PGM images to color PPM images
-    for (i = 0; i < h; i++) {
-        for (j = 0; j < w; j++) {
-            imgPPM.img_r[i * w + j] = data[i * w + j];
-            imgPPM.img_g[i * w + j] = data[i * w + j];
-            imgPPM.img_b[i * w + j] = data[i * w + j];
-        }
-    }
-
-    for (it = kpt_list.begin(); it != kpt_list.end(); it++) {
-        // derive circle radius cR
-        cR = (int)it->scale;
-        if (cR <= 1) { // avoid zero radius
-            cR = 1;
-        }
-        r = (int)it->r;
-        c = (int)it->c;
-        //  draw_red_circle(&imgPPM, r, c, cR);
-        rasterCircle(&imgPPM, r, c, cR);
-        rasterCircle(&imgPPM, r, c, cR + 1);
-        float ori = it->ori;
-        draw_red_orientation(&imgPPM, c, r, ori, cR);
-    }
-
-    // write rendered image to output
-    write_rgb2ppm(out_filename, imgPPM.img_r, imgPPM.img_g, imgPPM.img_b, w, h);
-
-    // free allocated memory
-    delete[] imgPPM.img_r;
-    delete[] imgPPM.img_g;
-    delete[] imgPPM.img_b;
-} // render()
-
-int export_kpt_list_to_file(const char *filename,
-                            std::list<SiftKeypoint> &kpt_list,
-                            bool bIncludeDescpritor)
-{
-    FILE *fp;
-    fp = fopen(filename, "wb");
-    if (!fp) {
-        printf("Fail to open file: %s\n", filename);
-        // exit(1);
-        return -1;
-    }
-
-    fprintf(fp, "%u\t%d\n", static_cast<unsigned int>(kpt_list.size()), 128);
-
-    std::list<SiftKeypoint>::iterator it;
-    for (it = kpt_list.begin(); it != kpt_list.end(); it++) {
-        fprintf(fp, "%d\t%d\t%.2f\t%.2f\t%.2f\t%.2f\t", it->octave, it->layer,
-                it->r, it->c, it->scale, it->ori);
-        if (bIncludeDescpritor) {
-            for (int i = 0; i < 128; i++) {
-                fprintf(fp, "%d\t", (int)(it->descriptors[i]));
-            }
-        }
-        fprintf(fp, "\n");
-    }
-
-    fclose(fp);
-    return 0;
-}
-
-// Draw a while line on a gray-scale image.
-// MatchPair mp indicates the coordinates of the start point
-// and the end point.
-int draw_line_to_image(Image<unsigned char> &image, MatchPair &mp)
-{
-    int w = image.w;
-    int r1 = mp.r1;
-    int c1 = mp.c1;
-    int r2 = mp.r2;
-    int c2 = mp.c2 + w / 2;
-
-    float k = (float)(r2 - r1) / (float)(c2 - c1);
-    for (int c = c1; c < c2; c++) {
-        // Line equation
-        int r = (int)(k * (c - c1) + r1);
-        image.data[r * w + c] = 190; // Draw a gray line.
-    }
-    return 0;
-}
-
-// Draw a line on the RGB color image.
-int draw_line_to_rgb_image(unsigned char *&data, int w, int h, MatchPair &mp)
-{
-    int r1 = mp.r1;
-    int c1 = mp.c1;
-    int r2 = mp.r2;
-    int c2 = mp.c2;
-
-    float k = (float)(r2 - r1) / (float)(c2 - c1);
-    for (int c = c1; c < c2; c++) {
-        // Line equation
-        int r = (int)(k * (c - c1) + r1);
-
-        // Draw a blue line
-        data[r * w * 3 + 3 * c] = 0;
-        data[r * w * 3 + 3 * c + 1] = 0;
-        data[r * w * 3 + 3 * c + 2] = 255;
-    }
-
-    return 0;
-}
-
-// Draw match lines between matched keypoints between two images.
-int draw_match_lines_to_ppm_file(const char *filename,
-                                 Image<unsigned char> &image1,
-                                 Image<unsigned char> &image2,
-                                 std::list<MatchPair> &match_list)
-{
-    Image<unsigned char> tmpImage;
-    combine_image(tmpImage, image1, image2);
-
-    int w = tmpImage.w;
-    int h = tmpImage.h;
-    unsigned char *srcData = tmpImage.data;
-    unsigned char *dstData = new unsigned char[w * h * 3];
-
-    for (int i = 0; i < w * h; i++) {
-        dstData[i * 3 + 0] = srcData[i];
-        dstData[i * 3 + 1] = srcData[i];
-        dstData[i * 3 + 2] = srcData[i];
-    }
-
-    std::list<MatchPair>::iterator p;
-    for (p = match_list.begin(); p != match_list.end(); p++) {
-        MatchPair mp;
-        mp.r1 = p->r1;
-        mp.c1 = p->c1;
-        mp.r2 = p->r2;
-        mp.c2 = p->c2 + image1.w;
-        draw_line_to_rgb_image(dstData, w, h, mp);
-    }
-
-    write_ppm(filename, dstData, w, h);
-
-    delete[] dstData;
     return 0;
 }
