+++ src/vvector.h
@@ -25,332 +25,304 @@ extern "C" {
 /* Zero out a 2D vector */
 
 #define VEC_ZERO_2(a)                                                          \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         (a)[0] = (a)[1] = 0.0;                                                 \
-    \
-}
+    }
 
 /* ========================================================== */
 /* Zero out a 3D vector */
 
 #define VEC_ZERO(a)                                                            \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         (a)[0] = (a)[1] = (a)[2] = 0.0;                                        \
-    \
-}
+    }
 
 /* ========================================================== */
 /* Zero out a 4D vector */
 
 #define VEC_ZERO_4(a)                                                          \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         (a)[0] = (a)[1] = (a)[2] = (a)[3] = 0.0;                               \
-    \
-}
+    }
 
 /* ========================================================== */
 /* Vector copy */
 
 #define VEC_COPY_2(b, a)                                                       \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         (b)[0] = (a)[0];                                                       \
         (b)[1] = (a)[1];                                                       \
-    \
-}
+    }
 
 /* ========================================================== */
 /* Copy 3D vector */
 
 #define VEC_COPY(b, a)                                                         \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         (b)[0] = (a)[0];                                                       \
         (b)[1] = (a)[1];                                                       \
         (b)[2] = (a)[2];                                                       \
-    \
-}
+    }
 
 /* ========================================================== */
 /* Copy 4D vector */
 
 #define VEC_COPY_4(b, a)                                                       \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         (b)[0] = (a)[0];                                                       \
         (b)[1] = (a)[1];                                                       \
         (b)[2] = (a)[2];                                                       \
         (b)[3] = (a)[3];                                                       \
-    \
-}
+    }
 
 /* ========================================================== */
 /* Vector difference */
 
 #define VEC_DIFF_2(v21, v2, v1)                                                \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         (v21)[0] = (v2)[0] - (v1)[0];                                          \
         (v21)[1] = (v2)[1] - (v1)[1];                                          \
-    \
-}
+    }
 
 /* ========================================================== */
 /* Vector difference */
 
 #define VEC_DIFF(v21, v2, v1)                                                  \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         (v21)[0] = (v2)[0] - (v1)[0];                                          \
         (v21)[1] = (v2)[1] - (v1)[1];                                          \
         (v21)[2] = (v2)[2] - (v1)[2];                                          \
-    \
-}
+    }
 
 /* ========================================================== */
 /* Vector difference */
 
 #define VEC_DIFF_4(v21, v2, v1)                                                \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         (v21)[0] = (v2)[0] - (v1)[0];                                          \
         (v21)[1] = (v2)[1] - (v1)[1];                                          \
         (v21)[2] = (v2)[2] - (v1)[2];                                          \
         (v21)[3] = (v2)[3] - (v1)[3];                                          \
-    \
-}
+    }
 
 /* ========================================================== */
 /* Vector sum */
 
 #define VEC_SUM_2(v21, v2, v1)                                                 \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         (v21)[0] = (v2)[0] + (v1)[0];                                          \
         (v21)[1] = (v2)[1] + (v1)[1];                                          \
-    \
-}
+    }
 
 /* ========================================================== */
 /* Vector sum */
 
 #define VEC_SUM(v21, v2, v1)                                                   \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         (v21)[0] = (v2)[0] + (v1)[0];                                          \
         (v21)[1] = (v2)[1] + (v1)[1];                                          \
         (v21)[2] = (v2)[2] + (v1)[2];                                          \
-    \
-}
+    }
 
 /* ========================================================== */
 /* Vector sum */
 
 #define VEC_SUM_4(v21, v2, v1)                                                 \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         (v21)[0] = (v2)[0] + (v1)[0];                                          \
         (v21)[1] = (v2)[1] + (v1)[1];                                          \
         (v21)[2] = (v2)[2] + (v1)[2];                                          \
         (v21)[3] = (v2)[3] + (v1)[3];                                          \
-    \
-}
+    }
 
 /* ========================================================== */
 /* scalar times vector */
 
 #define VEC_SCALE_2(c, a, b)                                                   \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         (c)[0] = (a) * (b)[0];                                                 \
         (c)[1] = (a) * (b)[1];                                                 \
-    \
-}
+    }
 
 /* ========================================================== */
 /* scalar times vector */
 
 #define VEC_SCALE(c, a, b)                                                     \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         (c)[0] = (a) * (b)[0];                                                 \
         (c)[1] = (a) * (b)[1];                                                 \
         (c)[2] = (a) * (b)[2];                                                 \
-    \
-}
+    }
 
 /* ========================================================== */
 /* scalar times vector */
 
 #define VEC_SCALE_4(c, a, b)                                                   \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         (c)[0] = (a) * (b)[0];                                                 \
         (c)[1] = (a) * (b)[1];                                                 \
         (c)[2] = (a) * (b)[2];                                                 \
         (c)[3] = (a) * (b)[3];                                                 \
-    \
-}
+    }
 
 /* ========================================================== */
 /* accumulate scaled vector */
 
 #define VEC_ACCUM_2(c, a, b)                                                   \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         (c)[0] += (a) * (b)[0];                                                \
         (c)[1] += (a) * (b)[1];                                                \
-    \
-}
+    }
 
 /* ========================================================== */
 /* accumulate scaled vector */
 
 #define VEC_ACCUM(c, a, b)                                                     \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         (c)[0] += (a) * (b)[0];                                                \
         (c)[1] += (a) * (b)[1];                                                \
         (c)[2] += (a) * (b)[2];                                                \
-    \
-}
+    }
 
 /* ========================================================== */
 /* accumulate scaled vector */
 
 #define VEC_ACCUM_4(c, a, b)                                                   \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         (c)[0] += (a) * (b)[0];                                                \
         (c)[1] += (a) * (b)[1];                                                \
         (c)[2] += (a) * (b)[2];                                                \
         (c)[3] += (a) * (b)[3];                                                \
-    \
-}
+    }
 
 /* ========================================================== */
 /* Vector dot product */
 
 #define VEC_DOT_PRODUCT_2(c, a, b)                                             \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         c = (a)[0] * (b)[0] + (a)[1] * (b)[1];                                 \
-    \
-}
+    }
 
 /* ========================================================== */
 /* Vector dot product */
 
 #define VEC_DOT_PRODUCT_3(c, a, b)                                             \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         c = (a)[0] * (b)[0] + (a)[1] * (b)[1] + (a)[2] * (b)[2];               \
-    \
-}
+    }
 
 /* ========================================================== */
 /* Vector dot product */
 
 #define VEC_DOT_PRODUCT_4(c, a, b)                                             \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         c = (a)[0] * (b)[0] + (a)[1] * (b)[1] + (a)[2] * (b)[2] +              \
             (a)[3] * (b)[3];                                                   \
-    \
-}
+    }
 
 /* ========================================================== */
 /* vector impact parameter (squared) */
 
 #define VEC_IMPACT_SQ(bsq, direction, position)                                \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         gleDouble vlen, llel;                                                  \
         VEC_DOT_PRODUCT(vlen, position, position);                             \
         VEC_DOT_PRODUCT(llel, direction, position);                            \
         bsq = vlen - llel * llel;                                              \
-    \
-}
+    }
 
 /* ========================================================== */
 /* vector impact parameter */
 
 #define VEC_IMPACT(bsq, direction, position)                                   \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         VEC_IMPACT_SQ(bsq, direction, position);                               \
         bsq = sqrt(bsq);                                                       \
-    \
-}
+    }
 
 /* ========================================================== */
 /* Vector length */
 
 #define VEC_LENGTH_2(vlen, a)                                                  \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         vlen = a[0] * a[0] + a[1] * a[1];                                      \
         vlen = sqrt(vlen);                                                     \
-    \
-}
+    }
 
 /* ========================================================== */
 /* Vector length */
 
 #define VEC_LENGTH(vlen, a)                                                    \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         vlen = (a)[0] * (a)[0] + (a)[1] * (a)[1];                              \
         vlen += (a)[2] * (a)[2];                                               \
         vlen = sqrt(vlen);                                                     \
-    \
-}
+    }
 
 /* ========================================================== */
 /* Vector length */
 
 #define VEC_LENGTH_4(vlen, a)                                                  \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         vlen = (a)[0] * (a)[0] + (a)[1] * (a)[1];                              \
         vlen += (a)[2] * (a)[2];                                               \
         vlen += (a)[3] * (a)[3];                                               \
         vlen = sqrt(vlen);                                                     \
-    \
-}
+    }
 
 /* ========================================================== */
 /* distance between two points */
 
 #define VEC_DISTANCE(vlen, va, vb)                                             \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         gleDouble tmp[4];                                                      \
         VEC_DIFF(tmp, vb, va);                                                 \
         VEC_LENGTH(vlen, tmp);                                                 \
-    \
-}
+    }
 
 /* ========================================================== */
 /* Vector length */
 
 #define VEC_CONJUGATE_LENGTH(vlen, a)                                          \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         vlen = 1.0 - a[0] * a[0] - a[1] * a[1] - a[2] * a[2];                  \
         vlen = sqrt(vlen);                                                     \
-    \
-}
+    }
 
 /* ========================================================== */
 /* Vector length */
 
 #define VEC_NORMALIZE(a)                                                       \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         double vlen;                                                           \
         VEC_LENGTH(vlen, a);                                                   \
         if (vlen != 0.0) {                                                     \
@@ -359,15 +331,14 @@ extern "C" {
             a[1] *= vlen;                                                      \
             a[2] *= vlen;                                                      \
         }                                                                      \
-    \
-}
+    }
 
 /* ========================================================== */
 /* Vector length */
 
 #define VEC_RENORMALIZE(a, newlen)                                             \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         double vlen;                                                           \
         VEC_LENGTH(vlen, a);                                                   \
         if (vlen != 0.0) {                                                     \
@@ -376,120 +347,111 @@ extern "C" {
             a[1] *= vlen;                                                      \
             a[2] *= vlen;                                                      \
         }                                                                      \
-    \
-}
+    }
 
 /* ========================================================== */
 /* 3D Vector cross product yeilding vector */
 
 #define VEC_CROSS_PRODUCT(c, a, b)                                             \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         c[0] = (a)[1] * (b)[2] - (a)[2] * (b)[1];                              \
         c[1] = (a)[2] * (b)[0] - (a)[0] * (b)[2];                              \
         c[2] = (a)[0] * (b)[1] - (a)[1] * (b)[0];                              \
-    \
-}
+    }
 
 /* ========================================================== */
 /* Vector perp -- assumes that n is of unit length
  * accepts vector v, subtracts out any component parallel to n */
 
 #define VEC_PERP(vp, v, n)                                                     \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         double vdot;                                                           \
                                                                                \
         VEC_DOT_PRODUCT(vdot, v, n);                                           \
         vp[0] = (v)[0] - (vdot) * (n)[0];                                      \
         vp[1] = (v)[1] - (vdot) * (n)[1];                                      \
         vp[2] = (v)[2] - (vdot) * (n)[2];                                      \
-    \
-}
+    }
 
 /* ========================================================== */
 /* Vector parallel -- assumes that n is of unit length
  * accepts vector v, subtracts out any component perpendicular to n */
 
 #define VEC_PARALLEL(vp, v, n)                                                 \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         double vdot;                                                           \
                                                                                \
         VEC_DOT_PRODUCT(vdot, v, n);                                           \
         vp[0] = (vdot) * (n)[0];                                               \
         vp[1] = (vdot) * (n)[1];                                               \
         vp[2] = (vdot) * (n)[2];                                               \
-    \
-}
+    }
 
 /* ========================================================== */
 /* Vector reflection -- assumes n is of unit length */
 /* Takes vector v, reflects it against reflector n, and returns vr */
 
 #define VEC_REFLECT(vr, v, n)                                                  \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         double vdot;                                                           \
                                                                                \
         VEC_DOT_PRODUCT(vdot, v, n);                                           \
         vr[0] = (v)[0] - 2.0 * (vdot) * (n)[0];                                \
         vr[1] = (v)[1] - 2.0 * (vdot) * (n)[1];                                \
         vr[2] = (v)[2] - 2.0 * (vdot) * (n)[2];                                \
-    \
-}
+    }
 
 /* ========================================================== */
 /* Vector blending */
 /* Takes two vectors a, b, blends them together */
 
 #define VEC_BLEND(vr, sa, a, sb, b)                                            \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
                                                                                \
         vr[0] = (sa) * (a)[0] + (sb) * (b)[0];                                 \
         vr[1] = (sa) * (a)[1] + (sb) * (b)[1];                                 \
         vr[2] = (sa) * (a)[2] + (sb) * (b)[2];                                 \
-    \
-}
+    }
 
 /* ========================================================== */
 /* Vector print */
 
 #define VEC_PRINT_2(a)                                                         \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         double vlen;                                                           \
         VEC_LENGTH_2(vlen, a);                                                 \
         printf(" a is %f %f length of a is %f \n", a[0], a[1], vlen);          \
-    \
-}
+    }
 
 /* ========================================================== */
 /* Vector print */
 
 #define VEC_PRINT(a)                                                           \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         double vlen;                                                           \
         VEC_LENGTH(vlen, (a));                                                 \
         printf(" a is %f %f %f length of a is %f \n", (a)[0], (a)[1], (a)[2],  \
                vlen);                                                          \
-    \
-}
+    }
 
 /* ========================================================== */
 /* Vector print */
 
 #define VEC_PRINT_4(a)                                                         \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         double vlen;                                                           \
         VEC_LENGTH_4(vlen, (a));                                               \
         printf(" a is %f %f %f %f length of a is %f \n", (a)[0], (a)[1],       \
                (a)[2], (a)[3], vlen);                                          \
-    \
-}
+    }
 
 /* ========================================================== */
 /* print matrix */
@@ -509,8 +471,7 @@ extern "C" {
                 printf(" \n");                                                 \
             }                                                                  \
         }                                                                      \
-    \
-}
+    }
 
 /* ========================================================== */
 /* print matrix */
@@ -530,8 +491,7 @@ extern "C" {
                 printf(" \n");                                                 \
             }                                                                  \
         }                                                                      \
-    \
-}
+    }
 
 /* ========================================================== */
 /* print matrix */
@@ -551,15 +511,14 @@ extern "C" {
                 printf(" \n");                                                 \
             }                                                                  \
         }                                                                      \
-    \
-}
+    }
 
 /* ========================================================== */
 /* initialize matrix */
 
 #define IDENTIFY_MATRIX_3X3(m)                                                 \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         m[0][0] = 1.0;                                                         \
         m[0][1] = 0.0;                                                         \
         m[0][2] = 0.0;                                                         \
@@ -571,15 +530,14 @@ extern "C" {
         m[2][0] = 0.0;                                                         \
         m[2][1] = 0.0;                                                         \
         m[2][2] = 1.0;                                                         \
-    \
-}
+    }
 
 /* ========================================================== */
 /* initialize matrix */
 
 #define IDENTIFY_MATRIX_4X4(m)                                                 \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         m[0][0] = 1.0;                                                         \
         m[0][1] = 0.0;                                                         \
         m[0][2] = 0.0;                                                         \
@@ -599,29 +557,27 @@ extern "C" {
         m[3][1] = 0.0;                                                         \
         m[3][2] = 0.0;                                                         \
         m[3][3] = 1.0;                                                         \
-    \
-}
+    }
 
 /* ========================================================== */
 /* matrix copy */
 
 #define COPY_MATRIX_2X2(b, a)                                                  \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         b[0][0] = a[0][0];                                                     \
         b[0][1] = a[0][1];                                                     \
                                                                                \
         b[1][0] = a[1][0];                                                     \
         b[1][1] = a[1][1];                                                     \
-    \
-}
+    }
 
 /* ========================================================== */
 /* matrix copy */
 
 #define COPY_MATRIX_2X3(b, a)                                                  \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         b[0][0] = a[0][0];                                                     \
         b[0][1] = a[0][1];                                                     \
         b[0][2] = a[0][2];                                                     \
@@ -629,15 +585,14 @@ extern "C" {
         b[1][0] = a[1][0];                                                     \
         b[1][1] = a[1][1];                                                     \
         b[1][2] = a[1][2];                                                     \
-    \
-}
+    }
 
 /* ========================================================== */
 /* matrix copy */
 
 #define COPY_MATRIX_3X3(b, a)                                                  \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         b[0][0] = a[0][0];                                                     \
         b[0][1] = a[0][1];                                                     \
         b[0][2] = a[0][2];                                                     \
@@ -649,15 +604,14 @@ extern "C" {
         b[2][0] = a[2][0];                                                     \
         b[2][1] = a[2][1];                                                     \
         b[2][2] = a[2][2];                                                     \
-    \
-}
+    }
 
 /* ========================================================== */
 /* matrix copy */
 
 #define COPY_MATRIX_4X4(b, a)                                                  \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         b[0][0] = a[0][0];                                                     \
         b[0][1] = a[0][1];                                                     \
         b[0][2] = a[0][2];                                                     \
@@ -677,29 +631,27 @@ extern "C" {
         b[3][1] = a[3][1];                                                     \
         b[3][2] = a[3][2];                                                     \
         b[3][3] = a[3][3];                                                     \
-    \
-}
+    }
 
 /* ========================================================== */
 /* matrix transpose */
 
 #define TRANSPOSE_MATRIX_2X2(b, a)                                             \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         b[0][0] = a[0][0];                                                     \
         b[0][1] = a[1][0];                                                     \
                                                                                \
         b[1][0] = a[0][1];                                                     \
         b[1][1] = a[1][1];                                                     \
-    \
-}
+    }
 
 /* ========================================================== */
 /* matrix transpose */
 
 #define TRANSPOSE_MATRIX_3X3(b, a)                                             \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         b[0][0] = a[0][0];                                                     \
         b[0][1] = a[1][0];                                                     \
         b[0][2] = a[2][0];                                                     \
@@ -711,15 +663,14 @@ extern "C" {
         b[2][0] = a[0][2];                                                     \
         b[2][1] = a[1][2];                                                     \
         b[2][2] = a[2][2];                                                     \
-    \
-}
+    }
 
 /* ========================================================== */
 /* matrix transpose */
 
 #define TRANSPOSE_MATRIX_4X4(b, a)                                             \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         b[0][0] = a[0][0];                                                     \
         b[0][1] = a[1][0];                                                     \
         b[0][2] = a[2][0];                                                     \
@@ -739,29 +690,27 @@ extern "C" {
         b[3][1] = a[1][3];                                                     \
         b[3][2] = a[2][3];                                                     \
         b[3][3] = a[3][3];                                                     \
-    \
-}
+    }
 
 /* ========================================================== */
 /* multiply matrix by scalar */
 
 #define SCALE_MATRIX_2X2(b, s, a)                                              \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         b[0][0] = (s)*a[0][0];                                                 \
         b[0][1] = (s)*a[0][1];                                                 \
                                                                                \
         b[1][0] = (s)*a[1][0];                                                 \
         b[1][1] = (s)*a[1][1];                                                 \
-    \
-}
+    }
 
 /* ========================================================== */
 /* multiply matrix by scalar */
 
 #define SCALE_MATRIX_3X3(b, s, a)                                              \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         b[0][0] = (s)*a[0][0];                                                 \
         b[0][1] = (s)*a[0][1];                                                 \
         b[0][2] = (s)*a[0][2];                                                 \
@@ -773,15 +722,14 @@ extern "C" {
         b[2][0] = (s)*a[2][0];                                                 \
         b[2][1] = (s)*a[2][1];                                                 \
         b[2][2] = (s)*a[2][2];                                                 \
-    \
-}
+    }
 
 /* ========================================================== */
 /* multiply matrix by scalar */
 
 #define SCALE_MATRIX_4X4(b, s, a)                                              \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         b[0][0] = (s)*a[0][0];                                                 \
         b[0][1] = (s)*a[0][1];                                                 \
         b[0][2] = (s)*a[0][2];                                                 \
@@ -801,29 +749,27 @@ extern "C" {
         b[3][1] = s * a[3][1];                                                 \
         b[3][2] = s * a[3][2];                                                 \
         b[3][3] = s * a[3][3];                                                 \
-    \
-}
+    }
 
 /* ========================================================== */
 /* multiply matrix by scalar */
 
 #define ACCUM_SCALE_MATRIX_2X2(b, s, a)                                        \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         b[0][0] += (s)*a[0][0];                                                \
         b[0][1] += (s)*a[0][1];                                                \
                                                                                \
         b[1][0] += (s)*a[1][0];                                                \
         b[1][1] += (s)*a[1][1];                                                \
-    \
-}
+    }
 
 /* +========================================================== */
 /* multiply matrix by scalar */
 
 #define ACCUM_SCALE_MATRIX_3X3(b, s, a)                                        \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         b[0][0] += (s)*a[0][0];                                                \
         b[0][1] += (s)*a[0][1];                                                \
         b[0][2] += (s)*a[0][2];                                                \
@@ -835,15 +781,14 @@ extern "C" {
         b[2][0] += (s)*a[2][0];                                                \
         b[2][1] += (s)*a[2][1];                                                \
         b[2][2] += (s)*a[2][2];                                                \
-    \
-}
+    }
 
 /* +========================================================== */
 /* multiply matrix by scalar */
 
 #define ACCUM_SCALE_MATRIX_4X4(b, s, a)                                        \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         b[0][0] += (s)*a[0][0];                                                \
         b[0][1] += (s)*a[0][1];                                                \
         b[0][2] += (s)*a[0][2];                                                \
@@ -863,31 +808,29 @@ extern "C" {
         b[3][1] += (s)*a[3][1];                                                \
         b[3][2] += (s)*a[3][2];                                                \
         b[3][3] += (s)*a[3][3];                                                \
-    \
-}
+    }
 
 /* +========================================================== */
 /* matrix product */
 /* c[x][y] = a[x][0]*b[0][y]+a[x][1]*b[1][y]+a[x][2]*b[2][y]+a[x][3]*b[3][y];*/
 
 #define MATRIX_PRODUCT_2X2(c, a, b)                                            \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         c[0][0] = a[0][0] * b[0][0] + a[0][1] * b[1][0];                       \
         c[0][1] = a[0][0] * b[0][1] + a[0][1] * b[1][1];                       \
                                                                                \
         c[1][0] = a[1][0] * b[0][0] + a[1][1] * b[1][0];                       \
         c[1][1] = a[1][0] * b[0][1] + a[1][1] * b[1][1];                       \
-    \
-}
+    }
 
 /* ========================================================== */
 /* matrix product */
 /* c[x][y] = a[x][0]*b[0][y]+a[x][1]*b[1][y]+a[x][2]*b[2][y]+a[x][3]*b[3][y];*/
 
 #define MATRIX_PRODUCT_3X3(c, a, b)                                            \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         c[0][0] = a[0][0] * b[0][0] + a[0][1] * b[1][0] + a[0][2] * b[2][0];   \
         c[0][1] = a[0][0] * b[0][1] + a[0][1] * b[1][1] + a[0][2] * b[2][1];   \
         c[0][2] = a[0][0] * b[0][2] + a[0][1] * b[1][2] + a[0][2] * b[2][2];   \
@@ -899,16 +842,15 @@ extern "C" {
         c[2][0] = a[2][0] * b[0][0] + a[2][1] * b[1][0] + a[2][2] * b[2][0];   \
         c[2][1] = a[2][0] * b[0][1] + a[2][1] * b[1][1] + a[2][2] * b[2][1];   \
         c[2][2] = a[2][0] * b[0][2] + a[2][1] * b[1][2] + a[2][2] * b[2][2];   \
-    \
-}
+    }
 
 /* ========================================================== */
 /* matrix product */
 /* c[x][y] = a[x][0]*b[0][y]+a[x][1]*b[1][y]+a[x][2]*b[2][y]+a[x][3]*b[3][y];*/
 
 #define MATRIX_PRODUCT_4X4(c, a, b)                                            \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         c[0][0] = a[0][0] * b[0][0] + a[0][1] * b[1][0] + a[0][2] * b[2][0] +  \
                   a[0][3] * b[3][0];                                           \
         c[0][1] = a[0][0] * b[0][1] + a[0][1] * b[1][1] + a[0][2] * b[2][1] +  \
@@ -944,38 +886,35 @@ extern "C" {
                   a[3][3] * b[3][2];                                           \
         c[3][3] = a[3][0] * b[0][3] + a[3][1] * b[1][3] + a[3][2] * b[2][3] +  \
                   a[3][3] * b[3][3];                                           \
-    \
-}
+    }
 
 /* ========================================================== */
 /* matrix times vector */
 
 #define MAT_DOT_VEC_2X2(p, m, v)                                               \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         p[0] = m[0][0] * v[0] + m[0][1] * v[1];                                \
         p[1] = m[1][0] * v[0] + m[1][1] * v[1];                                \
-    \
-}
+    }
 
 /* ========================================================== */
 /* matrix times vector */
 
 #define MAT_DOT_VEC_3X3(p, m, v)                                               \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         p[0] = m[0][0] * v[0] + m[0][1] * v[1] + m[0][2] * v[2];               \
         p[1] = m[1][0] * v[0] + m[1][1] * v[1] + m[1][2] * v[2];               \
         p[2] = m[2][0] * v[0] + m[2][1] * v[1] + m[2][2] * v[2];               \
-    \
-}
+    }
 
 /* ========================================================== */
 /* matrix times vector */
 
 #define MAT_DOT_VEC_4X4(p, m, v)                                               \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         p[0] =                                                                 \
             m[0][0] * v[0] + m[0][1] * v[1] + m[0][2] * v[2] + m[0][3] * v[3]; \
         p[1] =                                                                 \
@@ -984,21 +923,19 @@ extern "C" {
             m[2][0] * v[0] + m[2][1] * v[1] + m[2][2] * v[2] + m[2][3] * v[3]; \
         p[3] =                                                                 \
             m[3][0] * v[0] + m[3][1] * v[1] + m[3][2] * v[2] + m[3][3] * v[3]; \
-    \
-}
+    }
 
 /* ========================================================== */
 /* vector transpose times matrix */
 /* p[j] = v[0]*m[0][j] + v[1]*m[1][j] + v[2]*m[2][j]; */
 
 #define VEC_DOT_MAT_3X3(p, v, m)                                               \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         p[0] = v[0] * m[0][0] + v[1] * m[1][0] + v[2] * m[2][0];               \
         p[1] = v[0] * m[0][1] + v[1] * m[1][1] + v[2] * m[2][1];               \
         p[2] = v[0] * m[0][2] + v[1] * m[1][2] + v[2] * m[2][2];               \
-    \
-}
+    }
 
 /* ========================================================== */
 /* affine matrix times vector */
@@ -1006,12 +943,11 @@ extern "C" {
  * entries representing a translation */
 
 #define MAT_DOT_VEC_2X3(p, m, v)                                               \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         p[0] = m[0][0] * v[0] + m[0][1] * v[1] + m[0][2];                      \
         p[1] = m[1][0] * v[0] + m[1][1] * v[1] + m[1][2];                      \
-    \
-}
+    }
 
 /* ========================================================== */
 /* inverse transpose of matrix times vector
@@ -1025,8 +961,8 @@ extern "C" {
  */
 
 #define INV_TRANSP_MAT_DOT_VEC_2X2(p, m, v)                                    \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         gleDouble det;                                                         \
                                                                                \
         det = m[0][0] * m[1][1] - m[0][1] * m[1][0];                           \
@@ -1039,8 +975,7 @@ extern "C" {
             p[0] *= det;                                                       \
             p[1] *= det;                                                       \
         }                                                                      \
-    \
-}
+    }
 
 /* ========================================================== */
 /* transform normal vector by inverse transpose of matrix
@@ -1052,8 +987,8 @@ extern "C" {
  */
 
 #define NORM_XFORM_2X2(p, m, v)                                                \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         double mlen;                                                           \
                                                                                \
         if ((m[0][1] != 0.0) || (m[1][0] != 0.0) || (m[0][0] != m[1][1])) {    \
@@ -1068,8 +1003,7 @@ extern "C" {
         else {                                                                 \
             VEC_COPY_2(p, v);                                                  \
         }                                                                      \
-    \
-}
+    }
 
 /* ========================================================== */
 /* outer product of vector times vector transpose
@@ -1079,15 +1013,14 @@ extern "C" {
  */
 
 #define OUTER_PRODUCT_2X2(m, v, t)                                             \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         m[0][0] = v[0] * t[0];                                                 \
         m[0][1] = v[0] * t[1];                                                 \
                                                                                \
         m[1][0] = v[1] * t[0];                                                 \
         m[1][1] = v[1] * t[1];                                                 \
-    \
-}
+    }
 
 /* ========================================================== */
 /* outer product of vector times vector transpose
@@ -1097,8 +1030,8 @@ extern "C" {
  */
 
 #define OUTER_PRODUCT_3X3(m, v, t)                                             \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         m[0][0] = v[0] * t[0];                                                 \
         m[0][1] = v[0] * t[1];                                                 \
         m[0][2] = v[0] * t[2];                                                 \
@@ -1110,8 +1043,7 @@ extern "C" {
         m[2][0] = v[2] * t[0];                                                 \
         m[2][1] = v[2] * t[1];                                                 \
         m[2][2] = v[2] * t[2];                                                 \
-    \
-}
+    }
 
 /* ========================================================== */
 /* outer product of vector times vector transpose
@@ -1121,8 +1053,8 @@ extern "C" {
  */
 
 #define OUTER_PRODUCT_4X4(m, v, t)                                             \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         m[0][0] = v[0] * t[0];                                                 \
         m[0][1] = v[0] * t[1];                                                 \
         m[0][2] = v[0] * t[2];                                                 \
@@ -1142,8 +1074,7 @@ extern "C" {
         m[3][1] = v[3] * t[1];                                                 \
         m[3][2] = v[3] * t[2];                                                 \
         m[3][3] = v[3] * t[3];                                                 \
-    \
-}
+    }
 
 /* +========================================================== */
 /* outer product of vector times vector transpose
@@ -1153,15 +1084,14 @@ extern "C" {
  */
 
 #define ACCUM_OUTER_PRODUCT_2X2(m, v, t)                                       \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         m[0][0] += v[0] * t[0];                                                \
         m[0][1] += v[0] * t[1];                                                \
                                                                                \
         m[1][0] += v[1] * t[0];                                                \
         m[1][1] += v[1] * t[1];                                                \
-    \
-}
+    }
 
 /* +========================================================== */
 /* outer product of vector times vector transpose
@@ -1171,8 +1101,8 @@ extern "C" {
  */
 
 #define ACCUM_OUTER_PRODUCT_3X3(m, v, t)                                       \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         m[0][0] += v[0] * t[0];                                                \
         m[0][1] += v[0] * t[1];                                                \
         m[0][2] += v[0] * t[2];                                                \
@@ -1184,8 +1114,7 @@ extern "C" {
         m[2][0] += v[2] * t[0];                                                \
         m[2][1] += v[2] * t[1];                                                \
         m[2][2] += v[2] * t[2];                                                \
-    \
-}
+    }
 
 /* +========================================================== */
 /* outer product of vector times vector transpose
@@ -1195,8 +1124,8 @@ extern "C" {
  */
 
 #define ACCUM_OUTER_PRODUCT_4X4(m, v, t)                                       \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         m[0][0] += v[0] * t[0];                                                \
         m[0][1] += v[0] * t[1];                                                \
         m[0][2] += v[0] * t[2];                                                \
@@ -1216,8 +1145,7 @@ extern "C" {
         m[3][1] += v[3] * t[1];                                                \
         m[3][2] += v[3] * t[2];                                                \
         m[3][3] += v[3] * t[3];                                                \
-    \
-}
+    }
 
 /* +========================================================== */
 /* determinant of matrix
@@ -1226,11 +1154,10 @@ extern "C" {
  */
 
 #define DETERMINANT_2X2(d, m)                                                  \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         d = m[0][0] * m[1][1] - m[0][1] * m[1][0];                             \
-    \
-}
+    }
 
 /* ========================================================== */
 /* determinant of matrix
@@ -1239,13 +1166,12 @@ extern "C" {
  */
 
 #define DETERMINANT_3X3(d, m)                                                  \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         d = m[0][0] * (m[1][1] * m[2][2] - m[1][2] * m[2][1]);                 \
         d -= m[0][1] * (m[1][0] * m[2][2] - m[1][2] * m[2][0]);                \
         d += m[0][2] * (m[1][0] * m[2][1] - m[1][1] * m[2][0]);                \
-    \
-}
+    }
 
 /* ========================================================== */
 /* i,j,th cofactor of a 4x4 matrix
@@ -1253,8 +1179,8 @@ extern "C" {
  */
 
 #define COFACTOR_4X4_IJ(fac, m, i, j)                                          \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         int ii[4], jj[4], k;                                                   \
                                                                                \
         /* compute which row, columnt to skip */                               \
@@ -1279,8 +1205,7 @@ extern "C" {
         if (k != (k / 2) * 2) {                                                \
             (fac) = -(fac);                                                    \
         }                                                                      \
-    \
-}
+    }
 
 /* ========================================================== */
 /* determinant of matrix
@@ -1289,8 +1214,8 @@ extern "C" {
  */
 
 #define DETERMINANT_4X4(d, m)                                                  \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         double cofac;                                                          \
         COFACTOR_4X4_IJ(cofac, m, 0, 0);                                       \
         d = m[0][0] * cofac;                                                   \
@@ -1300,8 +1225,7 @@ extern "C" {
         d += m[0][2] * cofac;                                                  \
         COFACTOR_4X4_IJ(cofac, m, 0, 3);                                       \
         d += m[0][3] * cofac;                                                  \
-    \
-}
+    }
 
 /* ========================================================== */
 /* cofactor of matrix
@@ -1310,14 +1234,13 @@ extern "C" {
  */
 
 #define COFACTOR_2X2(a, m)                                                     \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         a[0][0] = (m)[1][1];                                                   \
         a[0][1] = -(m)[1][0];                                                  \
         a[1][0] = -(m)[0][1];                                                  \
         a[1][1] = (m)[0][0];                                                   \
-    \
-}
+    }
 
 /* ========================================================== */
 /* cofactor of matrix
@@ -1326,8 +1249,8 @@ extern "C" {
  */
 
 #define COFACTOR_3X3(a, m)                                                     \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         a[0][0] = m[1][1] * m[2][2] - m[1][2] * m[2][1];                       \
         a[0][1] = -(m[1][0] * m[2][2] - m[2][0] * m[1][2]);                    \
         a[0][2] = m[1][0] * m[2][1] - m[1][1] * m[2][0];                       \
@@ -1337,8 +1260,7 @@ extern "C" {
         a[2][0] = m[0][1] * m[1][2] - m[0][2] * m[1][1];                       \
         a[2][1] = -(m[0][0] * m[1][2] - m[0][2] * m[1][0]);                    \
    a[2][2] = m[0][0]*m[1][1] - m[0][1]*m[1][0]);                               \
-    \
-}
+    }
 
 /* ========================================================== */
 /* cofactor of matrix
@@ -1347,8 +1269,8 @@ extern "C" {
  */
 
 #define COFACTOR_4X4(a, m)                                                     \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         int i, j;                                                              \
                                                                                \
         for (i = 0; i < 4; i++) {                                              \
@@ -1356,8 +1278,7 @@ extern "C" {
                 COFACTOR_4X4_IJ(a[i][j], m, i, j);                             \
             }                                                                  \
         }                                                                      \
-    \
-}
+    }
 
 /* ========================================================== */
 /* adjoint of matrix
@@ -1367,14 +1288,13 @@ extern "C" {
  */
 
 #define ADJOINT_2X2(a, m)                                                      \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         a[0][0] = (m)[1][1];                                                   \
         a[1][0] = -(m)[1][0];                                                  \
         a[0][1] = -(m)[0][1];                                                  \
         a[1][1] = (m)[0][0];                                                   \
-    \
-}
+    }
 
 /* ========================================================== */
 /* adjoint of matrix
@@ -1384,8 +1304,8 @@ extern "C" {
  */
 
 #define ADJOINT_3X3(a, m)                                                      \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         a[0][0] = m[1][1] * m[2][2] - m[1][2] * m[2][1];                       \
         a[1][0] = -(m[1][0] * m[2][2] - m[2][0] * m[1][2]);                    \
         a[2][0] = m[1][0] * m[2][1] - m[1][1] * m[2][0];                       \
@@ -1395,8 +1315,7 @@ extern "C" {
         a[0][2] = m[0][1] * m[1][2] - m[0][2] * m[1][1];                       \
         a[1][2] = -(m[0][0] * m[1][2] - m[0][2] * m[1][0]);                    \
    a[2][2] = m[0][0]*m[1][1] - m[0][1]*m[1][0]);                               \
-    \
-}
+    }
 
 /* ========================================================== */
 /* adjoint of matrix
@@ -1406,8 +1325,8 @@ extern "C" {
  */
 
 #define ADJOINT_4X4(a, m)                                                      \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         int i, j;                                                              \
                                                                                \
         for (i = 0; i < 4; i++) {                                              \
@@ -1415,8 +1334,7 @@ extern "C" {
                 COFACTOR_4X4_IJ(a[j][i], m, i, j);                             \
             }                                                                  \
         }                                                                      \
-    \
-}
+    }
 
 /* ========================================================== */
 /* compute adjoint of matrix and scale
@@ -1425,14 +1343,13 @@ extern "C" {
  */
 
 #define SCALE_ADJOINT_2X2(a, s, m)                                             \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         a[0][0] = (s)*m[1][1];                                                 \
         a[1][0] = -(s)*m[1][0];                                                \
         a[0][1] = -(s)*m[0][1];                                                \
         a[1][1] = (s)*m[0][0];                                                 \
-    \
-}
+    }
 
 /* ========================================================== */
 /* compute adjoint of matrix and scale
@@ -1441,8 +1358,8 @@ extern "C" {
  */
 
 #define SCALE_ADJOINT_3X3(a, s, m)                                             \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         a[0][0] = (s) * (m[1][1] * m[2][2] - m[1][2] * m[2][1]);               \
         a[1][0] = (s) * (m[1][2] * m[2][0] - m[1][0] * m[2][2]);               \
         a[2][0] = (s) * (m[1][0] * m[2][1] - m[1][1] * m[2][0]);               \
@@ -1454,8 +1371,7 @@ extern "C" {
         a[0][2] = (s) * (m[0][1] * m[1][2] - m[0][2] * m[1][1]);               \
         a[1][2] = (s) * (m[0][2] * m[1][0] - m[0][0] * m[1][2]);               \
         a[2][2] = (s) * (m[0][0] * m[1][1] - m[0][1] * m[1][0]);               \
-    \
-}
+    }
 
 /* ========================================================== */
 /* compute adjoint of matrix and scale
@@ -1464,8 +1380,8 @@ extern "C" {
  */
 
 #define SCALE_ADJOINT_4X4(a, s, m)                                             \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         int i, j;                                                              \
                                                                                \
         for (i = 0; i < 4; i++) {                                              \
@@ -1474,8 +1390,7 @@ extern "C" {
                 a[j][i] *= s;                                                  \
             }                                                                  \
         }                                                                      \
-    \
-}
+    }
 
 /* ========================================================== */
 /* inverse of matrix
@@ -1485,14 +1400,13 @@ extern "C" {
  */
 
 #define INVERT_2X2(b, det, a)                                                  \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         double tmp;                                                            \
         DETERMINANT_2X2(det, a);                                               \
         tmp = 1.0 / (det);                                                     \
         SCALE_ADJOINT_2X2(b, tmp, a);                                          \
-    \
-}
+    }
 
 /* ========================================================== */
 /* inverse of matrix
@@ -1502,14 +1416,13 @@ extern "C" {
  */
 
 #define INVERT_3X3(b, det, a)                                                  \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         float tmp;                                                             \
         DETERMINANT_3X3(det, a);                                               \
         tmp = 1.0f / (det);                                                    \
         SCALE_ADJOINT_3X3(b, tmp, a);                                          \
-    \
-}
+    }
 
 /* ========================================================== */
 /* inverse of matrix
@@ -1519,14 +1432,13 @@ extern "C" {
  */
 
 #define INVERT_4X4(b, det, a)                                                  \
-    \
-{                                                                         \
+                                                                               \
+    {                                                                          \
         double tmp;                                                            \
         DETERMINANT_4X4(det, a);                                               \
         tmp = 1.0 / (det);                                                     \
         SCALE_ADJOINT_4X4(b, tmp, a);                                          \
-    \
-}
+    }
 
 /* ========================================================== */
 #if defined(__cplusplus) || defined(c_plusplus)
